Changed applicable shorts(uint16_t) to bytes(uint8_t)
Found some low hanging memory saving fruit in the form of some
unnecessary 16bit ints.
Could probably also change the mapObject x and y to uint8_t's however
can't be sure that maps won't be too big in the future

+++ TheBounceArduboy/Player.cpp
@@ -91,7 +91,7 @@ void PlayerClass::update(Arduboy& arduboy)
 	y += yVelocity;
 	x += xVelocity;
 }
-void PlayerClass::respawn(short spawnX, short spawnY)
+void PlayerClass::respawn(byte spawnX, byte spawnY)
 {
 	x = spawnX;
 	y = spawnY;
